backend/competitor_agent/requirements.txt
fastapi
uvicorn
redis
requests
python-dotenv